web/requirements.txt
flask == 2.2.5
Werkzeug==2.2.2
flask-sqlalchemy
flask_wtf
wtforms
email-validator
flask_bcrypt
flask_login
pyjwt
psycopg2
gunicorn